+++ CHANGELOG.md
@@ -2,12 +2,11 @@
 
 ## Unreleased
 
+- feat!: update to ESP-IDF v5 and new I2S API in hal
 - refactor!: do not set a default for num_frames_per_write
-
 - fix: use new name of library in task name
 - feat: make pinned_core_id public so Backend struct can be instantiated
 - feat: add report-render-time cargo feature
-
 - refactor: switch buffer from u8 to i16
 
 ## 0.2.0 - 2023-05-11